+++ components/agent/training.tsx
@@ -1,5 +1,5 @@
 'use client';
-import { useCompany } from '@/components/idiot/auth/hooks/useUser';
+import { useCompany } from '@/components/idiot/auth/useUser';
 import { Alert, AlertDescription } from '@/components/ui/alert';
 import { Button } from '@/components/ui/button';
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
