+++ components/idiot/auth/Profile.tsx
@@ -1,11 +1,15 @@
 'use client';
-
 import axios from 'axios';
 import { deleteCookie, getCookie } from 'cookies-next';
+import { ReactNode, useState } from 'react';
+import useSWR from 'swr';
+import { useRouter } from 'next/navigation';
+import { useAuthentication } from '@/components/idiot/auth/Router';
+import { DynamicFormFieldValueTypes } from '@/components/idiot/dynamic-form/DynamicForm';
+import { Button } from '@/components/ui/button';
 import { mutate } from 'swr';
 import DynamicForm from '@/components/idiot/dynamic-form/DynamicForm';
 import { Separator } from '@/components/ui/separator';
-import { Button } from '@/components/ui/button';
 import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert';
 import Link from 'next/link';
 
@@ -152,3 +156,71 @@ export const Profile = ({
     </div>
   );
 };
+
+export type ManageProps = {
+  userDataSWRKey?: string;
+  userDataEndpoint?: string;
+  userUpdateEndpoint?: string;
+  userPasswordChangeEndpoint?: string;
+};
+
+export default function Manage({
+  userDataSWRKey = '/user',
+  userDataEndpoint = '/v1/user',
+  userUpdateEndpoint = '/v1/user',
+}: ManageProps): ReactNode {
+  const [responseMessage, setResponseMessage] = useState('');
+  type User = {
+    missing_requirements?: {
+      [key: string]: {
+        type: 'number' | 'boolean' | 'text' | 'password';
+        value: DynamicFormFieldValueTypes;
+        validation?: (value: DynamicFormFieldValueTypes) => boolean;
+      };
+    };
+  };
+  const router = useRouter();
+  const authConfig = useAuthentication();
+  const { data, error, isLoading } = useSWR<User, any, string>(userDataSWRKey, async () => {
+    return (
+      await axios.get(`${authConfig.authServer}${userDataEndpoint}`, {
+        headers: {
+          'Content-Type': 'application/json',
+          Authorization: getCookie('jwt'),
+        },
+        validateStatus: (status) => [200, 403].includes(status),
+      })
+    ).data;
+  });
+
+  return (
+    <div className='w-full'>
+      <main className='flex min-h-[calc(100vh_-_theme(spacing.16))] flex-1 flex-col gap-4 bg-transparent p-4 md:gap-8 md:p-10'>
+        <div className='flex justify-between w-full max-w-6xl gap-2 mx-auto'>
+          {authConfig.manage.heading && <h2 className='text-3xl font-semibold'>{authConfig.manage.heading}</h2>}
+          <Button
+            key='done'
+            onClick={() => {
+              router.push('/chat');
+            }}
+          >
+            Go to {authConfig.appName}
+          </Button>
+        </div>
+        <Profile
+          {...{
+            isLoading,
+            error,
+            data,
+            router,
+            authConfig,
+            userDataSWRKey,
+            responseMessage,
+            userUpdateEndpoint,
+            setResponseMessage,
+          }}
+        />
+      </main>
+    </div>
+  );
+}
